feat: working on adding menu items to cart, got it working except you have to submit the request twice.

+++ backend/src/main/java/com/carolinawings/webapp/model/MenuItemOption.java
@@ -3,6 +3,8 @@ package com.carolinawings.webapp.model;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
@@ -14,7 +16,7 @@ public class MenuItemOption {
     private Long id;
 
     private String name; // sauce/dressing/other name
-    private Long price; //price of sauce if not
+    private BigDecimal price; //price of sauce if not
     @ManyToOne
     private OptionGroup optionGroup;
